Split out export formatting from KeepExportController
